+++ app/src/main/res/xml/preferences_general.xml
@@ -44,47 +44,4 @@
 
     </PreferenceCategory>
 
-    <PreferenceCategory
-        android:title="@string/settings_anime"
-        app:iconSpaceReserved="false">
-
-        <SwitchPreference
-            android:defaultValue="false"
-            android:key="@string/settings_anime_my_ongoings_key"
-            android:summaryOff="@string/settings_anime_my_ongoings_summary_off"
-            android:summaryOn="@string/settings_anime_my_ongoings_summary_on"
-            android:title="@string/settings_anime_my_ongoings_title"
-            app:iconSpaceReserved="false" />
-
-        <Preference
-            android:defaultValue="fandub"
-            android:key="@string/settings_anime_translation_type_key"
-            android:summary="@string/translation_voice"
-            android:title="@string/settings_anime_translation_type_title"
-            app:iconSpaceReserved="false" />
-
-        <Preference
-            android:key="@string/settings_anime_player_type_key"
-            android:summary="@string/settings_anime_player_type_summary"
-            android:title="@string/settings_anime_player_type_title"
-            app:iconSpaceReserved="false" />
-
-        <SwitchPreference
-            android:defaultValue="false"
-            android:key="@string/settings_anime_player_remember_key"
-            android:summaryOff="@string/settings_anime_player_remember_summary_off"
-            android:summaryOn="@string/settings_anime_player_remember_summary_on"
-            android:title="@string/settings_anime_player_remember_title"
-            app:iconSpaceReserved="false" />
-
-        <SwitchPreference
-            android:defaultValue="true"
-            android:key="@string/settings_anime_remember_translation_settings_key"
-            android:summaryOff="@string/settings_anime_remember_translation_settings_summary_off"
-            android:summaryOn="@string/settings_anime_remember_translation_settings_summary_on"
-            android:title="@string/settings_anime_remember_translation_settings_title"
-            app:iconSpaceReserved="false" />
-
-    </PreferenceCategory>
-
 </PreferenceScreen>
