--- app/src/main/res/xml/preferences_player.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <SwitchPreference
-        android:defaultValue="true"
-        android:key="@string/settings_player_orientation_key"
-        android:summaryOff="@string/settings_player_orientation_summary_off"
-        android:summaryOn="@string/settings_player_orientation_summary_on"
-        android:title="@string/settings_player_orientation_title"
-        app:iconSpaceReserved="false" />
-
-    <PreferenceCategory
-        android:title="@string/settings_player_gestures"
-        app:iconSpaceReserved="false">
-
-        <SwitchPreference
-            android:defaultValue="true"
-            android:key="@string/settings_player_gestures_key"
-            android:summaryOff="@string/settings_player_gestures_summary_off"
-            android:summaryOn="@string/settings_player_gestures_summary_on"
-            android:title="@string/settings_player_gestures_title"
-            app:iconSpaceReserved="false" />
-
-        <SwitchPreference
-            android:defaultValue="false"
-            android:dependency="@string/settings_player_gestures_key"
-            android:disableDependentsState="true"
-            android:key="@string/settings_player_gestures_offset_key"
-            android:summaryOff="@string/settings_player_gestures_offset_summary_off"
-            android:summaryOn="@string/settings_player_gestures_offset_summary_on"
-            android:title="@string/settings_player_gestures_offset_title"
-            app:iconSpaceReserved="false" />
-
-        <SwitchPreference
-            android:defaultValue="true"
-            android:dependency="@string/settings_player_gestures_offset_key"
-            android:key="@string/settings_player_gestures_controls_key"
-            android:summaryOff="@string/settings_player_gestures_controls_summary_off"
-            android:summaryOn="@string/settings_player_gestures_controls_summary_on"
-            android:title="@string/settings_player_gestures_controls_title"
-            app:iconSpaceReserved="false" />
-
-        <SwitchPreference
-            android:defaultValue="false"
-            android:dependency="@string/settings_player_gestures_controls_key"
-            android:key="@string/settings_player_gestures_controls_inverted_key"
-            android:summaryOff="@string/settings_player_gestures_controls_inverted_summary_off"
-            android:summaryOn="@string/settings_player_gestures_controls_inverted_summary_on"
-            android:title="@string/settings_player_gestures_controls_inverted_title"
-            app:iconSpaceReserved="false" />
-
-        <Preference
-            android:dependency="@string/settings_player_gestures_offset_key"
-            android:key="@string/settings_player_gestures_offset_small_key"
-            android:summary="@string/settings_player_gestures_offset_small_summary_default"
-            android:title="@string/settings_player_gestures_offset_small_title"
-            app:iconSpaceReserved="false" />
-
-        <Preference
-            android:dependency="@string/settings_player_gestures_offset_key"
-            android:key="@string/settings_player_gestures_offset_big_key"
-            android:summary="@string/settings_player_gestures_offset_big_summary_default"
-            android:title="@string/settings_player_gestures_offset_big_title"
-            app:iconSpaceReserved="false" />
-
-    </PreferenceCategory>
-
-</PreferenceScreen>
